third_party/WebKit/LayoutTests/editing/deleting/merge_into_empty_block.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
test(() => assert_selection(
    '<div contenteditable><div><br></div><blockquote>|foo</blockquote></div>',
    'delete',
    '<div contenteditable><blockquote>|foo</blockquote></div>'),
    'Hit backspace at beginning of BLOCKQUOTE');

test(() => assert_selection(
    '<div contenteditable><div><br></div><ul><li>|foo</li></ul></div>',
    'delete',
    '<div contenteditable><ul><li>|foo</li></ul></div>'),
    'Hit backspace at beginning of LI');
</script>
